tests/wpt/web-platform-tests/media-source/URL-createObjectURL-revoke.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Revoking a created URL with URL.revokeObjectURL(url)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
test(function() {
    var mediaSource = new MediaSource();
    var url = window.URL.createObjectURL(mediaSource);
    window.URL.revokeObjectURL(url);
    video.src = url;
    assert_equals(mediaSource.readyState, 'closed');
}, "Check revoking behavior of URL.revokeObjectURL(url).");
</script>
</body>
</html>
